deployment/requirements.txt
streamlit
pandas
scikit-learn
joblib
huggingface_hub
python-dotenv
